MQ-44: refactor signin request
